requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
django==4.0.6
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
pillow==9.2.0
psycopg2==2.9.3
pyjwt==2.4.0; python_version >= '3.6'
pytz==2022.1
setuptools==63.1.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
whitenoise==6.2.0
